Ein bisschen aufgeraeumt
Glove ist jetzt das Skript, wo die Klasse Glove und die Klasse TrackingData drin ist. Die Firewall ist konfiguriert und das UDPSend bekommt die Ethernet2 Addresse mitgeteilt.
