+++ Glove_Server_App/Assets/Code_Max/Glove.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3f3e01e70c3c7b343a1d5588a2b65990
+guid: e66ef96c12a386443a0f266d9797fc4a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
