+++ Glove_Server_App/Assets/serial_port_receiver.cs
@@ -11,65 +11,6 @@ using System.Net.Sockets;
 using System.Threading;
 using System.IO;
 
-static class Constants
-{
-	public const int NB_SENSORS = 40;
-	public const bool IS_BLUETOOTH = false;
-}
-//=======================================================
-//
-//=======================================================
-public class Glove {
-	public UInt16 NB_SENSORS = 40;
-	public UInt32 cnt;
-	public float[] values;
-	public UInt16 version;    
-
-	private Int64[] raw_values;
-	private Int64[] offsets;
-
-	public Glove(){
-		cnt = 0;
-		version = 0;
-		raw_values = new Int64[Constants.NB_SENSORS];
-		offsets = new Int64[Constants.NB_SENSORS];
-		values = new float[Constants.NB_SENSORS];
-	}
-
-	public void set_zero(	){
-		for(int i =0;i<Constants.NB_SENSORS;i++){
-			offsets[i] = raw_values[i];
-		}
-	}
-
-	public void apply_packet(Packet packet){
-		version = packet.version;
-		cnt++;
-
-		for (int i = 0; i < Constants.NB_SENSORS; i++) {
-			raw_values [i] = (Int64)(raw_values [i] + packet.values [i]);
-		}
-		raw_values [packet.key] = packet.value;
-
-		for (int i = 0; i < Constants.NB_SENSORS; i++) {
-			values [i] = 0.001f * (raw_values [i] - offsets [i]);
-		}
-		//Debug.Log ("cnt " + cnt);
-	}
-
-    public TrackingData GetTrackingData()
-    {
-        Vector3 vel = new Vector3(1, 2, 3);
-        Vector3 acc = new Vector3(2, 4, 6);
-        Matrix4x4 pose = new Matrix4x4(new Vector4(1, 0, 0, 0), new Vector4(0, 1, 0, 0), new Vector4(0, 0, 1, 0), new Vector4(0, 0, 0, 1));
-
-        return new TrackingData(values, pose, vel, acc, 2.0);
-    }
-
-}
-//=======================================================
-//
-//=======================================================
 public class Packet {
 
 	public  Char marker;
@@ -85,7 +26,6 @@ public class Packet {
 	public Packet(){
 		values = new Int16[Constants.NB_SENSORS];
 	}
-
 }
 
 //=======================================================
